Organizes Code for Easier Unit Testing # Description - Moves source files into sep folder - Helper functions exported in sep code files
