+++ src/summary/helper.ts
@@ -1,0 +1,4 @@
+// TODO Why is this needed
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const checkSummaryTemplateValid = (_summaryTemplate: string): boolean =>
+  true;
