+++ src/utils/index.ts
@@ -1,2 +1,6 @@
-export * as HelperUtils from './helper';
+export * as CountUtils from './count.utils';
+export * as DateTimeUtils from './date-time.utils';
+export * as DomUtils from './dom.utils';
+export * as NumberUtils from './number.utils';
 export * as ObjectUtils from './object.utils';
+export * as StringUtils from './string.utils';
